requirements.txt
loguru ~= 0.7
pytest ~= 8.1
pytest-cov ~= 5.0
python-dotenv ~= 1.0
pyyaml ~= 6.0
requests ~= 2.31
sqlalchemy ~= 2.0
pytest==8.1.1
pytest-cov==5.0.0
Werkzeug==3.0.1
Jinja2==3.1.3
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-crontab==3.0.0
python-dotenv==1.0.1